+++ flutter_inappwebview_forge_ios/ios/flutter_inappwebview_forge_ios/Sources/flutter_inappwebview_forge_ios/WebAuthenticationSession/WebAuthenticationSession.swift
@@ -39,6 +39,9 @@ public class WebAuthenticationSession: NSObject, ASWebAuthenticationPresentation
     public func prepare() {
         if let session = session as? ASWebAuthenticationSession {
             session.prefersEphemeralWebBrowserSession = settings.prefersEphemeralWebBrowserSession
+            if #available(iOS 17.4, *), let additionalHeaderFields = settings.additionalHeaderFields {
+                session.additionalHeaderFields = additionalHeaderFields
+            }
         }
     }