+++ flutter_inappwebview_forge_ios/ios/flutter_inappwebview_forge_ios/Sources/flutter_inappwebview_forge_ios/WebAuthenticationSession/WebAuthenticationSessionSettings.swift
@@ -12,6 +12,7 @@ import AuthenticationServices
 public class WebAuthenticationSessionSettings: ISettings<WebAuthenticationSession> {
     
     var prefersEphemeralWebBrowserSession = false
+    var additionalHeaderFields: [String: String]? = nil
     
     override init(){
         super.init()
@@ -21,6 +22,9 @@ public class WebAuthenticationSessionSettings: ISettings<WebAuthenticationSessio
         var realOptions: [String: Any?] = toMap()
         if let session = obj?.session as? ASWebAuthenticationSession {
             realOptions["prefersEphemeralWebBrowserSession"] = session.prefersEphemeralWebBrowserSession
+            if #available(iOS 17.4, *) {
+                realOptions["additionalHeaderFields"] = session.additionalHeaderFields
+            }
         }
         return realOptions
     }